Move buildElasticQuery/Body to bodybuilder
