+++ lib/elasticsearch/elastic-body-builder.js
@@ -1,6 +1,8 @@
-const { EXCLUDE_FIELDS, ITEM_FILTER_AGGREGATIONS } = require('./config')
+const { EXCLUDE_FIELDS, ITEM_FILTER_AGGREGATIONS, SORT_FIELDS } = require('./config')
 const { deepValue } = require('../util')
 const { esRangeValue } = require('../utils/resource-helpers')
+const ApiRequest = require('../api-request')
+const ElasticQueryBuilder = require('../elasticsearch/elastic-query-builder')
 
 /**
  * Given a ES search body, returns same object modified to include the
@@ -319,11 +321,59 @@ const bodyForFindByUri = async function (recapBarcodesByStatus, params) {
   return body
 }
 
+/**
+ *  Given GET params, returns a plainobject suitable for use in a ES query.
+ *
+ * @param {object} params - A hash of request params including `filters`,
+ * `search_scope`, `q`
+ *
+ * @return {object} ES query object suitable to be POST'd to ES endpoint
+ */
+const buildElasticQuery = function (params) {
+  const request = ApiRequest.fromParams(params)
+
+  const builder = ElasticQueryBuilder.forApiRequest(request)
+  return builder.query.toJson()
+}
+
+/**
+ *  Given GET params, returns a plainobject with `from`, `size`, `query`,
+ *  `sort`, and any other params necessary to perform the ES query based
+ *  on the GET params.
+ *
+ *  @return {object} An object that can be posted directly to ES
+ */
+const buildElasticBody = function (params) {
+  const body = {
+    from: (params.per_page * (params.page - 1)),
+    size: params.per_page
+  }
+
+  body.query = buildElasticQuery(params)
+
+  // Apply sort:
+  let direction
+  let field
+
+  if (params.sort === 'relevance') {
+    field = '_score'
+    direction = 'desc'
+  } else {
+    field = SORT_FIELDS[params.sort].field || params.sort
+    direction = params.sort_direction || SORT_FIELDS[params.sort].initialDirection
+  }
+  body.sort = [{ [field]: direction }, { uri: 'asc' }]
+
+  return body
+}
+
 module.exports = {
   bodyForFindByUri,
   addInnerHits,
   itemsFilterContext,
   recapStatuses,
   itemStatusFilterWithUnavailableRecapItems,
-  itemsQueryContext
+  itemsQueryContext,
+  buildElasticQuery,
+  buildElasticBody
 }
