+++ lib/resources.js
@@ -15,13 +15,18 @@ const AvailableDeliveryLocationTypes = require('./available_delivery_location_ty
 
 const { parseParams } = require('../lib/util')
 
-const ApiRequest = require('./api-request')
-const ElasticQueryBuilder = require('./elasticsearch/elastic-query-builder')
-const { AGGREGATIONS_SPEC, EXCLUDE_FIELDS, SORT_FIELDS } = require('./elasticsearch/config')
+const { AGGREGATIONS_SPEC, EXCLUDE_FIELDS } = require('./elasticsearch/config')
 
 const errors = require('./errors')
 const { esRangeValue, parseSearchParams, nyplSourceAndId, itemsByFilter } = require('./utils/resource-helpers')
-const { bodyForFindByUri, addInnerHits, itemsFilterContext, itemsQueryContext } = require('./elasticsearch/elastic-body-builder')
+const {
+  bodyForFindByUri,
+  addInnerHits,
+  itemsFilterContext,
+  itemsQueryContext,
+  buildElasticQuery,
+  buildElasticBody
+} = require('./elasticsearch/elastic-body-builder')
 
 const RESOURCES_INDEX = process.env.RESOURCES_INDEX
 
@@ -358,50 +363,4 @@ module.exports = function (app, _private = null) {
     _private.aggregationQueriesForParams = aggregationQueriesForParams
     _private.mergeAggregationsResponses = mergeAggregationsResponses
   }
-}
-
-/**
- *  Given GET params, returns a plainobject with `from`, `size`, `query`,
- *  `sort`, and any other params necessary to perform the ES query based
- *  on the GET params.
- *
- *  @return {object} An object that can be posted directly to ES
- */
-const buildElasticBody = function (params) {
-  const body = {
-    from: (params.per_page * (params.page - 1)),
-    size: params.per_page
-  }
-
-  body.query = buildElasticQuery(params)
-
-  // Apply sort:
-  let direction
-  let field
-
-  if (params.sort === 'relevance') {
-    field = '_score'
-    direction = 'desc'
-  } else {
-    field = SORT_FIELDS[params.sort].field || params.sort
-    direction = params.sort_direction || SORT_FIELDS[params.sort].initialDirection
-  }
-  body.sort = [{ [field]: direction }, { uri: 'asc' }]
-
-  return body
-}
-
-/**
- *  Given GET params, returns a plainobject suitable for use in a ES query.
- *
- * @param {object} params - A hash of request params including `filters`,
- * `search_scope`, `q`
- *
- * @return {object} ES query object suitable to be POST'd to ES endpoint
- */
-const buildElasticQuery = function (params) {
-  const request = ApiRequest.fromParams(params)
-
-  const builder = ElasticQueryBuilder.forApiRequest(request)
-  return builder.query.toJson()
 }
